feat: #239 뷰분리
