+++ Spoony-iOS/Spoony-iOS/Source/Feature/Search/Views/SearchResultRow.swift
